requirements.txt
kagglehub==0.3.12
matplotlib==3.10.3
numpy==2.1.3
pandas==2.3.1
pillow==11.3.0
pygame==2.6.1
scikit-learn==1.7.1
tensorflow==2.19.0
